product Servces Controller done working on chatbot

+++ SpringEcom/src/main/java/com/telusko/SpringEcom/repo/OrderRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface OrderRepo extends JpaRepository<Order, Integer> {
+public interface OrderRepo extends JpaRepository<Order, Long> {
     Optional<Order> findByOrderId(String orderId);
 }
